refactor(playground-ui): move remaining components to ds folder
Move ScrollableContainer and Threads from src/components to src/ds/components
following the design system pattern with barrel exports.
