+++ packages/playground-ui/src/domains/mcps/components/mcp-table/mcp-table.tsx
@@ -7,7 +7,7 @@ import { Icon } from '@/ds/icons/Icon';
 import { ColumnDef, flexRender, getCoreRowModel, useReactTable } from '@tanstack/react-table';
 import React, { useState } from 'react';
 
-import { ScrollableContainer } from '@/components/scrollable-container';
+import { ScrollableContainer } from '@/ds/components/ScrollableContainer';
 import { Skeleton } from '@/ds/components/Skeleton';
 import { columns } from './columns';
