+++ packages/playground-ui/src/domains/tools/components/tool-table/tool-table.tsx
@@ -6,7 +6,7 @@ import { Icon } from '@/ds/icons/Icon';
 import { ColumnDef, flexRender, getCoreRowModel, useReactTable } from '@tanstack/react-table';
 import React, { useMemo, useState } from 'react';
 
-import { ScrollableContainer } from '@/components/scrollable-container';
+import { ScrollableContainer } from '@/ds/components/ScrollableContainer';
 import { Skeleton } from '@/ds/components/Skeleton';
 import { columns } from './columns';
 import { useLinkComponent } from '@/lib/framework';
