+++ packages/playground-ui/src/ds/components/ScrollableContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './scrollable-container';
