+++ packages/playground-ui/src/ds/components/Threads/index.ts
@@ -1,0 +1,1 @@
+export * from './threads';
