+++ packages/playground-ui/src/index.ts
@@ -6,7 +6,7 @@ export * from './domains/tools/index';
 export * from './domains/workflows/index';
 export * from './domains/templates/index';
 export * from './domains/observability/index';
-export * from './components/threads';
+export * from './ds/components/Threads';
 export * from './types';
 
 // DS Components - Existing
